Revert "Rename "integrator" to "solver""
This reverts commit 87d27d7d8f522f47750895b42c88b3a3f5ede6b3.
