+++ dynamiqs/integrators/core/fixed_step_stochastic_integrator.py
@@ -15,14 +15,14 @@ from ...qarrays.utils import stack
 from ...result import Result
 from ...utils.general import dag, expect
 from ...utils.operators import eye_like
-from .abstract_solver import StochasticBaseSolver
+from .abstract_integrator import StochasticBaseIntegrator
 from .interfaces import DSMEInterface, DSSEInterface, SolveInterface
-from .rouchon_solver import _cholesky_normalize
+from .rouchon_integrator import _cholesky_normalize
 from .save_mixin import DiffusiveSolveSaveMixin
 
 
 class DiffusiveState(eqx.Module):
-    """State for the diffusive SSE/SME fixed step solvers."""
+    """State for the diffusive SSE/SME fixed step integrators."""
 
     state: QArray  # state (integrated from initial to current time)
     Y: Array  # measurement (integrated from initial to current time)
@@ -45,10 +45,10 @@ def _is_linearly_spaced(
     return np.allclose(diffs, diffs[0], rtol=rtol, atol=atol)
 
 
-class DiffusiveSolveSolver(
-    StochasticBaseSolver, DiffusiveSolveSaveMixin, SolveInterface
+class DiffusiveSolveIntegrator(
+    StochasticBaseIntegrator, DiffusiveSolveSaveMixin, SolveInterface
 ):
-    """Solver solving the diffusive SSE/SME with a fixed step size solver."""
+    """Integrator solving the diffusive SSE/SME with a fixed step size integrator."""
 
     def __check_init__(self):
         # check that all tsave values are exact multiples of dt
@@ -188,14 +188,14 @@ class DiffusiveSolveSolver(
         pass
 
 
-class DSSEFixedStepSolver(DiffusiveSolveSolver, DSSEInterface):
+class DSSEFixedStepIntegrator(DiffusiveSolveIntegrator, DSSEInterface):
     @property
     def nmeas(self) -> int:
         return len(self.Ls)
 
 
-class DSSESolveEulerMayuramaSolver(DSSEFixedStepSolver):
-    """Solver solving the diffusive SSE with the Euler-Mayurama method."""
+class DSSESolveEulerMayuramaIntegrator(DSSEFixedStepIntegrator):
+    """Integrator solving the diffusive SSE with the Euler-Mayurama method."""
 
     def forward(self, t: Scalar, y: DiffusiveState, dW: Array) -> DiffusiveState:
         psi = y.state
@@ -229,17 +229,17 @@ class DSSESolveEulerMayuramaSolver(DSSEFixedStepSolver):
         return DiffusiveState(psi + dpsi, y.Y + dY)
 
 
-dssesolve_euler_maruyama_solver_constructor = DSSESolveEulerMayuramaSolver
+dssesolve_euler_maruyama_integrator_constructor = DSSESolveEulerMayuramaIntegrator
 
 
-class DSMEFixedStepSolver(DiffusiveSolveSolver, DSMEInterface):
+class DSMEFixedStepIntegrator(DiffusiveSolveIntegrator, DSMEInterface):
     @property
     def nmeas(self) -> int:
         return len(self.etas)
 
 
-class DSMESolveEulerMayuramaSolver(DSMEFixedStepSolver):
-    """Solver solving the diffusive SME with the Euler-Mayurama method."""
+class DSMESolveEulerMayuramaIntegrator(DSMEFixedStepIntegrator):
+    """Integrator solving the diffusive SME with the Euler-Mayurama method."""
 
     def forward(self, t: Scalar, y: DiffusiveState, dW: Array) -> DiffusiveState:
         # The diffusive SME for a single detector is:
@@ -254,7 +254,7 @@ class DSMESolveEulerMayuramaSolver(DSMEFixedStepSolver):
         LdL = sum([_L.dag() @ _L for _L in L])
 
         # === Lcal(rho)
-        # (see MEDiffraxSolver in `solvers/core/diffrax_solver.py`)
+        # (see MEDiffraxIntegrator in `integrators/core/diffrax_integrator.py`)
         Hnh = -1j * H - 0.5 * LdL
         tmp = Hnh @ rho + sum([0.5 * _L @ rho @ _L.dag() for _L in L])
         Lcal_rho = tmp + tmp.dag()
@@ -276,8 +276,8 @@ class DSMESolveEulerMayuramaSolver(DSMEFixedStepSolver):
         return DiffusiveState(rho + drho, y.Y + dY)
 
 
-class DSMESolveRouchon1Solver(DSMEFixedStepSolver, SolveInterface):
-    """Solver solving the diffusive SME with the Rouchon1 method."""
+class DSMESolveRouchon1Integrator(DSMEFixedStepIntegrator, SolveInterface):
+    """Integrator solving the diffusive SME with the Rouchon1 method."""
 
     def forward(self, t: Scalar, y: DiffusiveState, dW: Array) -> DiffusiveState:
         # The Rouchon update for a single loss channel is:
@@ -322,5 +322,5 @@ class DSMESolveRouchon1Solver(DSMEFixedStepSolver, SolveInterface):
         return DiffusiveState(rho, y.Y + dY)
 
 
-dsmesolve_euler_maruyama_solver_constructor = DSMESolveEulerMayuramaSolver
-dsmesolve_rouchon1_solver_constructor = DSMESolveRouchon1Solver
+dsmesolve_euler_maruyama_integrator_constructor = DSMESolveEulerMayuramaIntegrator
+dsmesolve_rouchon1_integrator_constructor = DSMESolveRouchon1Integrator
