+++ tests/floquet/floquet_qubit.py
@@ -8,10 +8,10 @@ from jaxtyping import PyTree
 
 from dynamiqs import basis, sigmam, sigmaz, stack
 from dynamiqs.gradient import Gradient
-from dynamiqs.method import Method
+from dynamiqs.integrators.apis.floquet import floquet
 from dynamiqs.options import Options
 from dynamiqs.result import FloquetResult
-from dynamiqs.solvers.apis.floquet import floquet
+from dynamiqs.method import Method
 from dynamiqs.time_qarray import CallableTimeQArray, modulated
 
 from ..system import System
