+++ tests/integrator_tester.py
@@ -14,7 +14,7 @@ from dynamiqs.options import Options
 from .system import System
 
 
-class SolverTester:
+class IntegratorTester:
     def _test_correctness(
         self,
         system: System,
