+++ tests/mesolve/test_adaptive.py
@@ -3,15 +3,15 @@ import pytest
 from dynamiqs.gradient import Autograd, CheckpointAutograd
 from dynamiqs.method import Tsit5
 
+from ..integrator_tester import IntegratorTester
 from ..order import TEST_LONG
-from ..solver_tester import SolverTester
 from .open_system import dense_ocavity, dia_ocavity, otdqubit
 
 # we only test Tsit5 to keep the unit test suite fast
 
 
 @pytest.mark.run(order=TEST_LONG)
-class TestMESolveAdaptive(SolverTester):
+class TestMESolveAdaptive(IntegratorTester):
     @pytest.mark.parametrize('system', [dense_ocavity, dia_ocavity, otdqubit])
     def test_correctness(self, system):
         self._test_correctness(system, Tsit5())
