+++ tests/mesolve/test_rouchon.py
@@ -3,13 +3,13 @@ import pytest
 from dynamiqs.gradient import Autograd, CheckpointAutograd
 from dynamiqs.method import Rouchon1
 
+from ..integrator_tester import IntegratorTester
 from ..order import TEST_LONG
-from ..solver_tester import SolverTester
 from .open_system import dense_ocavity, dia_ocavity, otdqubit
 
 
 @pytest.mark.run(order=TEST_LONG)
-class TestMESolveRouchon1(SolverTester):
+class TestMESolveRouchon1(IntegratorTester):
     @pytest.mark.parametrize('system', [dense_ocavity, dia_ocavity, otdqubit])
     @pytest.mark.parametrize('normalize', [True, False])
     def test_correctness(self, system, normalize):
