+++ tests/sesolve/test_euler.py
@@ -3,13 +3,13 @@ import pytest
 from dynamiqs.gradient import Autograd, CheckpointAutograd
 from dynamiqs.method import Euler
 
+from ..integrator_tester import IntegratorTester
 from ..order import TEST_LONG
-from ..solver_tester import SolverTester
 from .closed_system import dense_cavity, dia_cavity, tdqubit
 
 
 @pytest.mark.run(order=TEST_LONG)
-class TestSESolveEuler(SolverTester):
+class TestSESolveEuler(IntegratorTester):
     @pytest.mark.parametrize('system', [dense_cavity, dia_cavity, tdqubit])
     def test_correctness(self, system):
         method = Euler(dt=1e-4)
